moving and testing PiMAdParser
