+++ src/modularPlantManager/ModularPlantManager.ts
@@ -49,7 +49,7 @@ import {EventEmitter} from 'events';
 import StrictEventEmitter from 'strict-event-emitter-types';
 import PiMAdResponse = Backbone.PiMAdResponse;
 import { v4 as uuidv4 } from 'uuid';
-import {PiMAdParser} from './PiMAdParser';
+import {PiMAdParser} from './pea/PiMAdParser/PiMAdParser';
 import {ProcedureOptions} from '@p2olab/polaris-interface/dist/service/options';
